+++ Source/SwiftEntryKit.swift
@@ -27,27 +27,29 @@ public final class SwiftEntryKit {
     
     /**
      Displays a given entry view using an attributes struct.
-     - A thread-safe method - Can be invokes from any thread.
-     - A class method - Should be called on the class.
+     - A thread-safe method - Can be invokes from any thread
+     - A class method - Should be called on the class
      - parameter view: Custom view that is to be displayed
      - parameter attributes: Display properties
+     - parameter rollbackWindow: After the entry has been dismissed, SwiftEntryKit rolls back to the given window. By default it is the application key window
      */
-    public class func display(entry view: UIView, using attributes: EKAttributes) {
+    public class func display(entry view: UIView, using attributes: EKAttributes, rollbackWindow: UIWindow = UIApplication.shared.keyWindow!) {
         execute {
-            EKWindowProvider.shared.display(view: view, using: attributes)
+            EKWindowProvider.shared.display(view: view, using: attributes, rollbackWindow: rollbackWindow)
         }
     }
     
     /**
      Displays a given entry view controller using an attributes struct.
-     - A thread-safe method - Can be invokes from any thread.
-     - A class method - Should be called on the class.
+     - A thread-safe method - Can be invokes from any thread
+     - A class method - Should be called on the class
      - parameter view: Custom view that is to be displayed
      - parameter attributes: Display properties
+     - parameter rollbackWindow: After the entry has been dismissed, SwiftEntryKit rolls back to the given window. By default it is the application key window
      */
-    public class func display(entry viewController: UIViewController, using attributes: EKAttributes) {
+    public class func display(entry viewController: UIViewController, using attributes: EKAttributes, rollbackWindow: UIWindow = UIApplication.shared.keyWindow!) {
         execute {
-            EKWindowProvider.shared.display(viewController: viewController, using: attributes)
+            EKWindowProvider.shared.display(viewController: viewController, using: attributes, rollbackWindow: rollbackWindow)
         }
     }